🚧 BXQZ97 task: apply external agent result
