+++ .agentplane/tasks/202608261249-BXQZ97/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202608261249-BXQZ97/add-provider-update-branch-recovery",
   "created_at": "2026-08-26T13:12:22.036Z",
-  "diffstat_sha256": "sha256:7061390f48f6d9a0dc43f847789d7ff549efab36260534c493a90a31a820fd54",
+  "diffstat_sha256": "sha256:915fe26fe6d19af523588c1c59b9cccbea20ab5bb71010238aa110fc24122e2e",
   "last_verified_at": "2026-08-26T14:20:31.076Z",
   "last_verified_diffstat_sha256": "sha256:7061390f48f6d9a0dc43f847789d7ff549efab36260534c493a90a31a820fd54",
   "pr_number": 4892,
